application/Assassins/requirements.txt
amqp==1.4.6
anyjson==0.3.3
Babel==2.0
beautifulsoup4==4.4.0
billiard==3.3.0.20
bleach==1.4.1
celery==3.1.18
defusedxml==0.4.1
Django==1.8.4
django-admin-bootstrapped==2.5.4
django-allauth==0.23.0
django-appconf==1.0.1
django-commands==0.7
django-compressor==1.5
django-filter==0.11.0
django-haystack==2.4.0
django-modelcluster==0.6.2
-e git://github.com/rutledgepaulv/django-require.git@eed3c61989136c56d2492efd00345065cd117863#egg=django_require-master
django-taggit==0.17.0
django-treebeard==3.0
flower==0.8.3
gunicorn==19.3.0
html5lib==0.999999
kombu==3.0.26
oauthlib==1.0.3
Pillow==2.9.0
python3-openid==3.0.6
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
requirements==0.1
six==1.9.0
slimit==0.8.1
South==1.0.2
tornado==4.2.1
Unidecode==0.4.18
Whoosh==2.7.0
Willow==0.2.1
psycopg2==2.6.1